requirements.txt
opencv-python==4.6.0.66
imutils==0.5.4
grpcio==1.48.1
grpcio-tools==1.48.1
RPi.GPIO==0.7.1
python-dotenv==0.21.0
